// frontend/src/api/models/handlerWarehouseListResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerWarehouseListResponseStatus } from './handlerWarehouseListResponseStatus'
import type { HandlerWarehouseListResponseType } from './handlerWarehouseListResponseType'

/**
 * Warehouse list item with basic information
 */
export interface HandlerWarehouseListResponse {
  city?: string
  code?: string
  created_at?: string
  id?: string
  is_default?: boolean
  name?: string
  province?: string
  short_name?: string
  sort_order?: number
  status?: HandlerWarehouseListResponseStatus
  type?: HandlerWarehouseListResponseType
}