// frontend/src/api/models/handlerWarehouseListResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerWarehouseListResponseStatus =
  (typeof HandlerWarehouseListResponseStatus)[keyof typeof HandlerWarehouseListResponseStatus]

export const HandlerWarehouseListResponseStatus = {
  enabled: 'enabled',
  disabled: 'disabled',
} as const