// frontend/src/api/models/handlerWarehouseListResponseType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerWarehouseListResponseType =
  (typeof HandlerWarehouseListResponseType)[keyof typeof HandlerWarehouseListResponseType]

export const HandlerWarehouseListResponseType = {
  normal: 'normal',
  virtual: 'virtual',
  transit: 'transit',
} as const