// frontend/src/api/models/getPartnerWarehousesParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { GetPartnerWarehousesOrderDir } from './getPartnerWarehousesOrderDir'
import type { GetPartnerWarehousesStatus } from './getPartnerWarehousesStatus'
import type { GetPartnerWarehousesType } from './getPartnerWarehousesType'

export type GetPartnerWarehousesParams = {
  /**
   * Search term (name, code)
   */
  search?: string
  /**
   * Warehouse status
   */
  status?: GetPartnerWarehousesStatus
  /**
   * Warehouse type
   */
  type?: GetPartnerWarehousesType
  /**
   * City
   */
  city?: string
  /**
   * Province
   */
  province?: string
  /**
   * Filter by default status
   */
  is_default?: boolean
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Order by field
   */
  order_by?: string
  /**
   * Order direction
   */
  order_dir?: GetPartnerWarehousesOrderDir
}