// frontend/src/api/models/getPartnerWarehousesType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetPartnerWarehousesType =
  (typeof GetPartnerWarehousesType)[keyof typeof GetPartnerWarehousesType]

export const GetPartnerWarehousesType = {
  physical: 'physical',
  virtual: 'virtual',
  consign: 'consign',
  transit: 'transit',
} as const